Extract portfolio workflow mapper
