+++ tests/unit/test_portfolio_service.py
@@ -3,7 +3,6 @@ from typing import Any
 import pytest
 from fastapi import HTTPException
 
-from app.contracts.portfolio import PortfolioSummary, PortfolioWorkflowLaunchCue
 from app.services.portfolio_service import PortfolioService
 
 
@@ -1206,94 +1205,6 @@ async def test_portfolio_workflow_uses_latest_transaction_probe_for_activity_pre
             "reporting_currency": None,
         }
     ]
-
-
-def test_build_workflow_actions_dedupes_and_ignores_unsupported_cues():
-    service = PortfolioService(_StubLotusCoreQueryClient())
-
-    actions = service._build_workflow_actions(
-        portfolio_id="PF_1001",
-        summary=PortfolioSummary(
-            assets_under_management_base=1000.0,
-            invested_market_value_base=900.0,
-            cash_market_value_base=100.0,
-            cash_weight_pct=10.0,
-            position_count=2,
-            cash_balance_count=1,
-        ),
-        operations=None,
-        workflow_cues=[
-            PortfolioWorkflowLaunchCue(
-                key="holdings",
-                label="Holdings",
-                href="/portfolio?portfolioId=PF_1001#portfolio-drilldown",
-            ),
-            PortfolioWorkflowLaunchCue(
-                key="custom",
-                label="Custom",
-                href="/custom",
-            ),
-            PortfolioWorkflowLaunchCue(
-                key="performance",
-                label="Performance",
-                href="/performance?portfolioId=PF_1001",
-            ),
-            PortfolioWorkflowLaunchCue(
-                key="holdings",
-                label="Holdings",
-                href="/portfolio?portfolioId=PF_1001#portfolio-drilldown",
-            ),
-        ],
-        transaction_total=3,
-    )
-
-    assert [action.title for action in actions] == [
-        "Review performance",
-        "Review holdings",
-    ]
-    assert [action.cta_label for action in actions] == [
-        "Performance",
-        "Holdings",
-    ]
-    assert actions[0].recommended is True
-
-
-def test_build_workflow_actions_returns_empty_portfolio_setup_sequence():
-    service = PortfolioService(_StubLotusCoreQueryClient())
-
-    actions = service._build_workflow_actions(
-        portfolio_id="PF_EMPTY",
-        summary=PortfolioSummary(
-            assets_under_management_base=0.0,
-            invested_market_value_base=0.0,
-            cash_market_value_base=0.0,
-            cash_weight_pct=0.0,
-            position_count=0,
-            cash_balance_count=0,
-        ),
-        operations=None,
-        workflow_cues=[],
-        transaction_total=0,
-    )
-
-    assert [action.title for action in actions] == [
-        "Fund portfolio",
-        "Book first trade",
-        "Publish pricing",
-        "Review holdings",
-        "Open performance",
-    ]
-    assert [action.sequence for action in actions] == [1, 2, 3, 4, 5]
-    assert [action.target for action in actions] == [
-        "Target: cash funding and opening balance setup",
-        "Target: transaction entry and execution workflow",
-        "Target: pricing publication and valuation refresh",
-        "Target: holdings and allocation review",
-        "Target: performance workspace after valuation is available",
-    ]
-    assert actions[0].recommended is True
-    assert actions[0].href == "/workbench?portfolioId=PF_EMPTY"
-    assert actions[-1].href == "/performance?portfolioId=PF_EMPTY"
 
 
 @pytest.mark.asyncio
